Clean up Project draft primitives
